requirements.txt
awsebcli==3.19.3
bcrypt==3.2.0
blinker==1.4
botocore==1.19.63
cement==2.8.2
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.3
dnspython==2.1.0
docker==4.4.3
ebcli==4.0.18
email-validator==1.1.2
enum34==1.1.10
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
future==0.16.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
keyring==9.0
keyrings.alt==4.0.2
MarkupSafe==1.1.1
mysql==0.0.2
mysqlclient==2.0.3
pathspec==0.5.9
pycparser==2.20
PyMySQL==1.0.2
pypiwin32==223
python-dateutil==2.8.1
python-dotenv==0.14.0
pywin32==227
pywin32-ctypes==0.2.0
PyYAML==5.3.1
requests==2.24.0
semantic-version==2.5.0
six==1.15.0
SQLAlchemy==1.3.23
termcolor==1.1.0
urllib3==1.25.11
wcwidth==0.1.9
websocket-client==0.57.0
Werkzeug==1.0.1
WTForms==2.3.3
